Replaced 'ignore' utils with fast forwarding in integration tests
